Turns out I completely forgot about this
Updated Extended Crafting
Switched to using utility recipe read/write methods

+++ src/main/java/thelm/packagedexcrafting/recipe/IRecipeInfoCombination.java
@@ -16,7 +16,7 @@ public interface IRecipeInfoCombination extends IRecipeInfo {
 
 	ItemStack getOutput();
 
-	int getEnergyRequired();
+	long getEnergyRequired();
 
 	int getEnergyUsage();
